% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTpLot.R
\name{eQTpLot}
\alias{eQTpLot}
\title{An eQTpLot function}
\usage{
eQTpLot(
  GWAS.df,
  eQTL.df,
  Genes.df,
  gene,
  trait,
  sigpvalue_GWAS = 5e-08,
  sigpvalue_eQTL = 0.05,
  tissue = "all",
  range = 200,
  NESeQTLRange = c(NA, NA),
  ylima = NA,
  ylimd = NA,
  xlimd = NA,
  genometrackheight = 1.5,
  gbuild = "hg19",
  res = 300,
  hgt = 18,
  wi = 18,
  getplot = TRUE,
  saveplot = TRUE
)
}
\arguments{
\item{GWAS.df}{Dataframe, one row per SNP, with columns
CHR Chromosome for SNP (X coded numerical ly as 23)
POS Chromosomal position for each SNP, in base pairs
SNP Variant ID (such ws dbSNP ID "rs...". Note: Must be the same naming scheme as  used in the eQTL.df to ensure proper matching)
pvalue_GWAS pvalue for the SNP from GWAS analysis
beta_GWAS beta for SNP from GWAS analysis
Trait name of trait for which GWAS was run. This column must be present, with a value for every row, even if your GWAS.df contains data for only one trait.
(Note: GWAS.df can contain multiple traits (such as one might get from a PheWAS. Specifying the trait to be analyzed will filter only for GWAS entires for this trait)}

\item{eQTL.df}{Dataframe, one row per SNP, with columns
SNP Variant ID (such ws dbSNP ID "rs...". Note: Must be the same naming scheme as  used in the GWAS.df to ensure proper matching)
Gene Gene symbol/name to which the eQTL expression data refers (Note: gene symbol/name must match entries in Genes.df to ensure proper matching)
pvalue_eQTL pvalue for the SNP from eQTL analysis (such as one might download from the GTEx Portal)
NES_eQTL NES (normalized effect size) for the SNP from eQTL analysis
(Per GTEx: defined as the slope of the linear regression, and is computed as the effect of the alternative allele (ALT) relative to the reference allele (REF) in the human genome reference
(i.e., the eQTL effect allele is the ALT allele). NES are computed in a normalized space where magnitude has no direct biological interpretation.)
Tissue tissue type to which the eQTL pvalue/beta refer
(Note: eQTL.df can contain multiple tissue types. Specifying the tissue type to be analyzed will filter only for eQTL entires for this tissue type.
Alternatively, setting tissue type to "all" (the default setting) will automatically pick the smallest eQTL pvalue for each SNP across all tissues for a PanTissue analysis)}

\item{Genes.df}{Dataframe, one row per gene, with the following columns
(Note: eQTpLot automatically loads a default Genes.df database containing information for both genomic builds hg19 and hg38,
but you may wish to specify our own Genes.df dataframe if your gene of interest is not included in the default dataframe, or if your eQTL data uses a different gene naming scheme
(for example, Gencode ID instead of gene symbol)):
Gene Gene symbol/name for which the Coordinate data (below) refers to
(Note: gene symbol/name must match entries in eQTL.df to ensure proper matching)
CHR Chromosome the gene is on (X coded numerically as 23)
Start Chromosomal coordinate of start position (in basepairs) to use for gene (Note: this should be the smaller of the two values between Start and Stop)
Stop Chromosomal coordinate of end position (in basepairs) to use for gene (Note: this should be the larger of the two values between Start and Stop)
Build The genome build (either hg19 or hg38) for the location data -
the default Genes.df dataframe contains entries for both genome builds for each gene, and the script will select the appropriate entry based on the specified gbuild (default is hg19)).}

\item{gene}{name/symbol of gene to analyze, in quotes}

\item{trait}{name of trait to analyze, in quotes (from GWAS.df)}

\item{sigpvalue_GWAS}{GWAS pvalue significance threshold to use (this value will be used for a horizontal line in plot A, and to define GWAS significant/non-significant SNPs for plot C)}

\item{tissue}{tissue type, in quotes, for eQTL data to use (from eQTL.df, default setting is "all" for Pan-Tissue analysis)}

\item{range}{range, in kB, to extend analysis window on either side of gene boundry. Default is 200 kb}

\item{NESeQTLRange}{the maximum and minimum limits in the format c(min,max), to display for the NES value in eQTL.df.
The default setting will adjust the size limits automatically for your data, whereas specifying the limits
can keep them consistent between plots.}

\item{ylima}{used to manually adjust the y axis limit in plot A, if needed}

\item{ylimd}{used to manually adjust the y axis limit in plot D, if needed}

\item{xlimd}{used to manually adjust the x axis limit in plot D, if needed}

\item{genometrackheight}{used to set the height of the genome track panel (B), with default setting of 1.5.
Gene-dense regions may require more plotting space, whereas gene-poor regions may look better with less plotting space.}

\item{gbuild}{the genome build to use, in quotes, for fetching genomic information for panel B.
Default is "hg19" but can specify "hg38" if needed. This build should match the genome build used for "CHR" and "POS" in the GWAS.df}

\item{res}{resolution of the output plot image (default is 300 dpi)}

\item{hgt}{height of the output plot image (default is 12 inches)}

\item{wi}{width of the output plot image (default is 14 inches)}

\item{getplot}{default is TRUE. If set to false, script will not dsiplay the generated plot in the viewer}

\item{saveplot}{default is TRUE, script will save the generated plot in the working diretory with the name
"gene.trait.tissue.eQTL.png" using the supplied variables}

\item{sigpvalue_eQTLeQTL}{pvalue significance threshold to use (eQTL data with a pvalue larger than this threshold will be excluded from the analysis)}
}
\description{
This function creates an eQTpLot composite
}
\examples{
Saves to current directory
eQTpLot(Genes.df = Genes.df.example, GWAS.df = GWAS.df.example,
        eQTL.df = eQTL.df.example, gene = "SPATA7", trait = "eGFR",
        getplot=FALSE)
eQTpLot()
}
